custom_pc_odoo_17_v1/models/custom_pc.py
```python
from odoo import models, fields, api

class CustomPC(models.Model):
    _name = 'custom.pc'
    _description = 'Custom PC'

    pc_component = fields.Many2one('product.product', string='PC Component')
    total_cost = fields.Float(compute='_compute_total_cost', string='Total Cost')
    compatibility_status = fields.Boolean(compute='_check_compatibility', string='Compatibility Status')
    customization_options = fields.Text(string='Customization Options')

    @api.depends('pc_component')
    def _compute_total_cost(self):
        for record in self:
            record.total_cost = sum(component.price for component in record.pc_component)

    @api.depends('pc_component')
    def _check_compatibility(self):
        for record in self:
            # Add your compatibility checking logic here
            record.compatibility_status = True

    def customize_pc(self, customization_options):
        self.customization_options = customization_options
```