custom_pc_odoo_17_v1/tests/test_custom_pc.py
```python
from odoo.tests import common

class TestCustomPC(common.TransactionCase):

    def setUp(self):
        super(TestCustomPC, self).setUp()
        self.CustomPC = self.env['custom_pc']

    def test_01_check_product_catalog(self):
        self.assertEqual(self.CustomPC.product_catalog, 'Test Catalog')

    def test_02_check_dynamic_pricing(self):
        self.assertEqual(self.CustomPC.calculate_total_cost(), 1000)

    def test_03_check_compatibility(self):
        self.assertTrue(self.CustomPC.check_compatibility())

    def test_04_check_customization_options(self):
        self.assertEqual(self.CustomPC.customization_options, 'Test Options')

    def test_05_check_error_handling(self):
        with self.assertRaises(ValueError):
            self.CustomPC.check_compatibility()

    def test_06_check_integration(self):
        self.assertTrue(self.CustomPC.integrate_with_inventory())
        self.assertTrue(self.CustomPC.integrate_with_sales())

    def test_07_check_scalability(self):
        self.assertTrue(self.CustomPC.handle_large_catalog())

    def test_08_check_security(self):
        with self.assertRaises(PermissionError):
            self.CustomPC.access_unauthorized_data()
```
